requirements.txt
instauto
pandas
numpy
scikit-learn
xgboost
seaborn
matplotlib
plotly
bs4
requests
datetime
yfinance
openpyxlx
